+++ lib/main.dart
@@ -1,9 +1,6 @@
-import 'package:debug_panel_devtool/pages/debug_panel_page.dart';
-import 'package:debug_panel_devtool/src/utils/debug_utils.dart';
+import 'package:debug_panel_devtool/src/pages/debug_panel_page.dart';
 import 'package:devtools_extensions/devtools_extensions.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
 
 void main() async => runApp(const MyApp());
 
@@ -12,17 +9,8 @@ class MyApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return DevToolsExtension(
-      child: ChangeNotifierProvider(
-        create: (context) => variableDebugData,
-        child: DebugVariablesPanel(
-          onEvent: (name, params) {
-            if (kDebugMode) {
-              debugPrint('$name, $params');
-            }
-          },
-        ),
-      ),
+    return const DevToolsExtension(
+      child: DebugPanelPage(),
     );
   }
 }
