+++ lib/src/themes/flutter_flow_default_theme.dart
@@ -1,8 +1,8 @@
 import 'package:debug_panel_devtool/src/consts/theme_values.dart';
-import 'package:debug_panel_devtool/themes/flutter_flow_theme.dart';
+import 'package:debug_panel_devtool/src/themes/flutter_flow_theme.dart';
 import 'package:flutter/material.dart';
 
-export 'package:debug_panel_devtool/themes/flutter_flow_theme.dart'
+export 'package:debug_panel_devtool/src/themes/flutter_flow_theme.dart'
     show TextStyleHelper;
 
 extension AppThemeExtensions on BuildContext {
