+++ lib/src/widgets/filter_icon_button.dart
@@ -6,7 +6,7 @@ import 'package:debug_panel_devtool/src/widgets/flutter_flow_checkbox.dart';
 import 'package:debug_panel_devtool/src/widgets/flutter_flow_multiselect_dropdown.dart';
 import 'package:debug_panel_devtool/src/widgets/model_menu_button.dart';
 import 'package:debug_panel_devtool/src/widgets/styled_tooltip.dart';
-import 'package:debug_panel_devtool/themes/flutter_flow_default_theme.dart';
+import 'package:debug_panel_devtool/src/themes/flutter_flow_default_theme.dart';
 import 'package:debug_panel_proto/debug_panel_proto.dart';
 import 'package:flutter/material.dart';
