+++ lib/src/widgets/flutter_flow_checkbox.dart
@@ -1,6 +1,6 @@
 import 'package:debug_panel_devtool/src/consts/theme_values.dart';
 import 'package:debug_panel_devtool/src/widgets/styled_tooltip.dart';
-import 'package:debug_panel_devtool/themes/flutter_flow_default_theme.dart';
+import 'package:debug_panel_devtool/src/themes/flutter_flow_default_theme.dart';
 import 'package:flutter/material.dart';
 
 class CheckboxBooleanProperty extends StatefulWidget {
