+++ lib/src/widgets/flutter_flow_hover.dart
@@ -1,5 +1,5 @@
 import 'package:debug_panel_devtool/src/utils/ff_utils.dart';
-import 'package:debug_panel_devtool/themes/flutter_flow_default_theme.dart';
+import 'package:debug_panel_devtool/src/themes/flutter_flow_default_theme.dart';
 import 'package:flutter/material.dart';
 
 /// Wrap a widget with [FlutterFlowHoverWidget] to define hover behaviors.
