+++ lib/src/widgets/narrow_modal.dart
@@ -2,7 +2,7 @@ import 'package:debug_panel_devtool/src/consts/theme_values.dart';
 import 'package:debug_panel_devtool/src/utils/debug_utils.dart';
 import 'package:debug_panel_devtool/src/utils/ff_icons.dart';
 import 'package:debug_panel_devtool/src/utils/ff_utils.dart';
-import 'package:debug_panel_devtool/themes/flutter_flow_default_theme.dart';
+import 'package:debug_panel_devtool/src/themes/flutter_flow_default_theme.dart';
 import 'package:debug_panel_proto/debug_panel_proto.dart';
 import 'package:flutter/material.dart';
 import 'package:flutterflow_tree_view/flutterflow_tree_view.dart';
