+++ lib/src/widgets/panel_search.dart
@@ -1,7 +1,7 @@
 import 'package:debug_panel_devtool/src/consts/theme_values.dart';
 import 'package:debug_panel_devtool/src/utils/ff_icons.dart';
 import 'package:debug_panel_devtool/src/utils/ff_utils.dart';
-import 'package:debug_panel_devtool/themes/flutter_flow_default_theme.dart';
+import 'package:debug_panel_devtool/src/themes/flutter_flow_default_theme.dart';
 import 'package:flutter/material.dart';
 
 class PanelSearchWidget extends StatefulWidget {
