+++ lib/src/widgets/property_name.dart
@@ -1,7 +1,7 @@
 import 'package:auto_size_text/auto_size_text.dart';
 import 'package:debug_panel_devtool/src/consts/theme_values.dart';
 import 'package:debug_panel_devtool/src/widgets/styled_tooltip.dart';
-import 'package:debug_panel_devtool/themes/flutter_flow_default_theme.dart';
+import 'package:debug_panel_devtool/src/themes/flutter_flow_default_theme.dart';
 import 'package:flutter/material.dart';
 
 class PropertyNameWidget extends StatelessWidget {
